+++ TABP.Application/Commands/SpecialOffers/DeleteSpecialOffer/DeleteSpecialOfferCommandHandler.cs
@@ -1,22 +1,29 @@
 ﻿using MediatR;
 using TABP.DAL.Interfaces;
 using TABP.DAL.Interfaces.Repositories;
+using TABP.Domain.Exceptions;
 
 namespace TABP.Application.Commands.SpecialOffers.DeleteSpecialOffer;
 
 public class DeleteSpecialOfferCommandHandler : IRequestHandler<DeleteSpecialOfferCommand>
 {
     private readonly ISpecialOfferRepository _specialOfferRepository;
+    private readonly IRoomRepository _roomRepository;
     private readonly IUnitOfWork _unitOfWork;
 
-    public DeleteSpecialOfferCommandHandler(IUnitOfWork unitOfWork, ISpecialOfferRepository specialOfferRepository)
+    public DeleteSpecialOfferCommandHandler(IUnitOfWork unitOfWork, ISpecialOfferRepository specialOfferRepository,
+        IRoomRepository roomRepository)
     {
         _unitOfWork = unitOfWork;
         _specialOfferRepository = specialOfferRepository;
+        _roomRepository = roomRepository;
     }
 
     public async Task Handle(DeleteSpecialOfferCommand request, CancellationToken cancellationToken)
     {
+        if (!await _roomRepository.ExistsAsync(r => r.Id == request.RoomId))
+            throw new NotFoundException($"Room with id {request.RoomId} wasn't found.");
+
         await _specialOfferRepository.DeleteAsync(request.Id, request.RoomId);
 
         await _unitOfWork.SaveChangesAsync();
